+++ R/methods.R
@@ -46,7 +46,7 @@ setMethod("subset", signature(x = "propr"),
 
             if(is.character(select)){
 
-              select <- which(colnames(x@counts) %in% select)
+              select <- match(select, colnames(x@counts))
             }
 
             x@counts <- x@counts[subset, select, drop = FALSE]
@@ -161,101 +161,11 @@ setMethod("plot", signature(x = "propr", y = "missing"),
                                                      y = "y.val",
                                                      group = "group")) +
               ggplot2::geom_path(ggplot2::aes(colour = factor(df$group))) +
-              ggplot2::labs(x = "Expression *LR mRNA[1]",
-                            y = "Expression *LR mRNA[2]") +
+              ggplot2::labs(x = "*lr-transformed Abundance[1]",
+                            y = "*lr-transformed Abundance[2]") +
               ggplot2::coord_equal(ratio = 1) +
               ggthemes::theme_base() +
               ggplot2::theme(legend.position = "none") +
-              ggplot2::ggtitle(title)
-            plot(p)
-
-            return(p)
-          }
-)
-
-#' @rdname propr
-#' @section Methods (by generic):
-#' \code{image:} Method to plot \code{propr} object.
-#'
-#' @param cexRow Numeric. Size of x-axis label.
-#' @param cexCol Numeric. Size of y-axis label.
-# #' @param object An object of class \code{propr}.
-# #' @param title A character string. A title for the \code{propr} plot.
-#' @export
-setMethod("image", signature(x = "propr"),
-          function(x, cexRow = 10, cexCol = 10, title = "*LR Transformed Image"){
-
-            if(!requireNamespace("ggplot2", quietly = TRUE)){
-              stop("Uh oh! This plot method depends on ggplot2! ",
-                   "Try running: install.packages('ggplot2')")
-            }
-
-            if(length(x@pairs) == 0){
-
-              cat("Alert: Generating plot using all feature pairs.\n")
-              i.feat <- 1:nrow(x@matrix)
-
-            }else{
-
-              cat("Alert: Generating plot using indexed feature pairs.\n")
-              V <- x@pairs
-              coord <- indexToCoord(V, nrow(x@matrix))
-              i.feat <- sort(union(coord[[1]], coord[[2]]))
-            }
-
-            nfeat <- length(i.feat)
-            nsubj <- nrow(x@logratio)
-
-            # Prepare features for melting
-            feat <- vector("character", nfeat * nsubj)
-            featnames <- colnames(x@logratio)[i.feat]
-            if(is.null(featnames)){
-              featnames <- paste("Feature", i.feat)
-              cexCol <- 0
-            }
-
-            # Prepare subjects for melting
-            subj <- vector("character", nfeat * nsubj)
-            subjnames <- rownames(x@logratio)
-            if(is.null(subjnames)){
-              subjnames <- paste("Subject", 1:nsubj)
-              cexRow <- 0
-            }
-
-            # Melt *lr counts by feature
-            val <- vector("numeric", nfeat * nsubj)
-            for(i in 1:nfeat){
-
-              feat[((i-1)*nsubj + 1):((i-1)*nsubj + nsubj)] <- featnames[i]
-              subj[((i-1)*nsubj + 1):((i-1)*nsubj + nsubj)] <- subjnames
-              val[((i-1)*nsubj + 1):((i-1)*nsubj + nsubj)] <- scale(x@logratio[, i])
-            }
-
-            # Plot *lr for each subject
-            df <- data.frame("feature" = feat, "subject" = subj, "value" = val,
-                             stringsAsFactors = FALSE)
-            df$feature <- factor(df$feature, levels = unique(df$feature))
-            df$subject <- factor(df$subject, levels = unique(df$subject))
-            valMin <- floor(min(df$value))
-            valMax <- ceiling(max(df$value))
-            p <- ggplot2::ggplot(ggplot2::aes_string(x = "subject",
-                                                     y = "feature"), data = df) +
-              ggplot2::geom_tile(ggplot2::aes_string(fill = "value")) +
-              ggplot2::scale_fill_gradient2(name = "Scaled *LR Expression",
-                                            # low = "grey0", high = "grey70", mid = "grey35",
-                                            low = "yellow", high = "red", mid = "orange",
-                                            midpoint = 0,
-                                            limit = c(valMin, valMax),
-                                            breaks = seq(valMin, valMax)) +
-              ggplot2::theme(axis.title.x = ggplot2::element_blank(),
-                             axis.title.y = ggplot2::element_blank(),
-                             axis.text.x = ggplot2::element_text(size = cexRow),
-                             axis.text.y = ggplot2::element_text(size = cexCol),
-                             panel.grid.major = ggplot2::element_blank(),
-                             panel.border = ggplot2::element_blank(),
-                             panel.background = ggplot2::element_blank(),
-                             axis.ticks = ggplot2::element_blank(),
-                             legend.position = "bottom") +
               ggplot2::ggtitle(title)
             plot(p)
 
@@ -271,12 +181,17 @@ setMethod("image", signature(x = "propr"),
 # #' @param title A character string. A title for the \code{propr} plot.
 #' @param group A character or numeric vector. Supply feature groups for coloring.
 #'  Feature groups expected in the order they appear in \code{@@counts}.
-#' @importFrom stats as.dist as.dendrogram hclust order.dendrogram
+#' @importFrom stats as.dist as.dendrogram order.dendrogram
 #' @importFrom grDevices rainbow
 #' @export
 dendrogram <- function(object, title = "Proportional Clusters", group){
 
-  if(!requireNamespace("dendextend", quietly = TRUE)){
+  if(suppressWarnings(!requireNamespace("fastcluster", quietly = TRUE))){
+    stop("Uh oh! This plot method depends on fastcluster! ",
+         "Try running: install.packages('fastcluster')")
+  }
+
+  if(suppressWarnings(!requireNamespace("dendextend", quietly = TRUE))){
     stop("Uh oh! This plot method depends on dendextend! ",
          "Try running: install.packages('dendextend')")
   }
@@ -330,9 +245,10 @@ dendrogram <- function(object, title = "Proportional Clusters", group){
     stop("Matrix style not recognized.")
   }
 
-  dend <- as.dendrogram(hclust(dist))
+  hc <- fastcluster::hclust(dist)
+  dend <- as.dendrogram(hc)
   dendextend::labels_colors(dend) <- colorKey$color[order.dendrogram(dend)]
   plot(dend, main = title)
 
-  return(dend)
+  return(hc)
 }